Added support for arguments to the debug component
The normal arguments like rdfaEditorInit callback, toolbarOptions, and
editorOptions are now supported (passed through) on the debug component.
