+++ tests/dummy/app/templates/debug.hbs
@@ -1,2 +1,19 @@
-<Rdfa::RdfaEditorWithDebug/>
-
+<Rdfa::RdfaEditorWithDebug
+  @rdfaEditorInit={{this.rdfaEditorInit}}
+  @randomText={{this.randomText}}
+  @editorOptions={{hash
+    showToggleRdfaAnnotations="true"
+    showInsertButton=null
+    showRdfa="true"
+    showRdfaHighlight="true"
+    showRdfaHover="true"
+    showPaper="true"
+    showSidebar="true"
+    showToolbarBottom=null
+  }}
+  @toolbarOptions={{hash
+    showTextStyleButtons="true"
+    showListButtons="true"
+    showIndentButtons="true"
+  }}
+/>
